fix(dF): initialize Plone site fields
Due to us using the plone.dexterity fork from Asko where we remove the
slow `__getattr__` behavior.

Here we run into a chicken and egg problem: the dexterity system and
portal types are not set up because we are creating the Plone Site
object, yet, we would need the portal types to know which fields to
initialize.